small changes for consistency

+++ src/app/api/event/route.ts
@@ -35,16 +35,16 @@ export const POST = async (request: NextRequest) => {
 export const PATCH = async (request: NextRequest) => {
   try {
     const { event } = await request.json();
-    
-      const updatedEvent = await prisma.event.update({
-        where: {
-          id: event.id,
-        },
-        data: { 
-          ...event, 
-          id: undefined, // needed to prevent prisma update error: overwriting "id"
-        },
-      });
+
+    const updatedEvent = await prisma.event.update({
+      where: {
+        id: event.id,
+      },
+      data: {
+        ...event,
+        id: undefined,
+      },
+    });
     return NextResponse.json(
       {
         code: "SUCCESS",
@@ -65,19 +65,30 @@ export const PATCH = async (request: NextRequest) => {
 };
 
 export const DELETE = async (request: NextRequest) => {
-  try {
-    const { id } = await request.json();
+  const { searchParams } = new URL(request.url);
+  const id = searchParams.get("id");
 
-    const deleteEvent = await prisma.event.delete({
-      where: {
-        id: id,
+  // Check if id is null
+  if (!id) {
+    return NextResponse.json(
+      {
+        code: "BAD_REQUEST",
+        message: "Event ID is required.",
       },
+      { status: 400 }
+    );
+  }
+
+  try {
+    const deletedEvent = await prisma.event.delete({
+      where: { id },
     });
 
     return NextResponse.json(
       {
         code: "SUCCESS",
-        message: deleteEvent.eventName,
+        message: "Event deleted successfully",
+        data: deletedEvent,
       },
       { status: 200 }
     );
@@ -93,56 +104,38 @@ export const DELETE = async (request: NextRequest) => {
   }
 };
 
-// NOTE: Changed to fetch only one event
 export const GET = async (request: NextRequest) => {
-  // currently implemented for fetching all events
-  // need to add filtering
   const { searchParams } = new URL(request.url);
-  const eventId: string | undefined = searchParams.get("id") || undefined;
+  const eventId = searchParams.get("id");
 
   try {
-    // const fetchedEvents = await prisma.event.findMany(); // Use findMany to fetch all events
-
     if (eventId) {
       const fetchedEvent = await prisma.event.findUnique({
-        where: {id: eventId}
+        where: { id: eventId },
       });
-  
 
       if (!fetchedEvent) {
         return NextResponse.json(
-          {
-            code: "NOT_FOUND",
-            message: "No event found",
-          },
+          { code: "NOT_FOUND", message: "No event found" },
           { status: 404 }
         );
       }
 
       return NextResponse.json(
-      {
-        code: "SUCCESS",
-        data: fetchedEvent,
-      },
-      { status: 200 }
+        { code: "SUCCESS", data: fetchedEvent },
+        { status: 200 }
       );
     } else {
+      const fetchedEvents = await prisma.event.findMany();
       return NextResponse.json(
-        {
-          code: "ERROR",
-          message: "No event ID provided",
-        },
-        { status: 400 }
+        { code: "SUCCESS", data: fetchedEvents },
+        { status: 200 }
       );
     }
-    
   } catch (error) {
     console.error("Error fetching events:", error);
     return NextResponse.json(
-      {
-        code: "ERROR",
-        message: "An error occurred while fetching events",
-      },
+      { code: "ERROR", message: "An error occurred while fetching events" },
       { status: 500 }
     );
   }
